Restore unified app launcher and fix used_quantity type validation
- Restored run_app.py to start both backend and frontend together
- Fixed ResponseValidationError by changing used_quantity from int to float
- Backend now correctly handles decimal values from database

+++ backend_gateway/routers/pantry_router.py
@@ -73,7 +73,7 @@ class UserPantryItem(BaseModel):
     unit_price: Optional[float]
     total_price: Optional[float]
     pantry_item_created_at: Optional[datetime]
-    used_quantity: Optional[int]
+    used_quantity: Optional[float]
     status: Optional[str]
     product_id: Optional[int]
     product_name: Optional[str]
